* BugFix: Corrections to the Unix file IO to allow for compilation on ARM Unix devics.
