+++ Engine/source/platformMac/macVolume.mm
@@ -176,8 +176,3 @@ Torque::FS::FileSystemRef Platform::FS::createNativeFS( const String &volume )
 {
    return new MacFileSystem( volume );
 }
-
-bool Torque::FS::VerifyWriteAccess(const Torque::Path &path)
-{
-   return true;
-}
